Renames the typo of the model name.
